+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * feat: Build QR code new UI (!148)
 * feat: Add wearables UI accesible from ORMS (!359)
 * feat: add patient demographics API endpoint (!309)
+* feat: unified Manage Caregivers module/page to manage the access to caregivers of patients (!492)
 
 ### Documentation